requirements.txt
coverage==6.3
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.2.0
django-bootstrap4==24.4
django-crispy-forms==2.3
git-filter-repo==2.45.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-decouple==3.8
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
whitenoise==6.8.2
pillow==10.2.0
django-storages==1.14.0
boto3==1.36.16
